Moved the fill_in_address test helper to lib/ for use by gem users.
