+++ spec/support/admin_addresses_helper.rb
@@ -196,56 +196,6 @@ module AdminAddresses
     [uri.path, uri.query].map{|s| s if s.present? }.compact.join('?')
   end
 
-  # Fill in an already loaded address form with the given +values+ (either a
-  # Spree::Address, or a Hash mapping field names to field values).  If +type+
-  # is :user, :bill, or :ship, then the address type field (for order
-  # addresses) will be set to User, Billing, or Shipping, respectively (see
-  # #assign_order_address in Spree::Admin::AddressesController).  The +type+
-  # parameter only works for admin UI address forms, not frontend.
-  #
-  # Note that JavaScript is required to fill in the state field.
-  def fill_in_address(values, type=nil)
-    if values.is_a?(Spree::Address)
-      fill_in Spree.t(:first_name), with: values.firstname
-      fill_in Spree.t(:last_name), with: values.lastname
-      fill_in Spree.t(:company), with: values.company if Spree::Config[:company]
-
-      if page.has_content?(/#{Regexp.escape(Spree.t(:street_address_2))}/i)
-        fill_in Spree.t(:street_address), with: values.address1
-        fill_in Spree.t(:street_address_2), with: values.address2
-      else
-        fill_in Spree.t(:address1), with: values.address1
-        fill_in Spree.t(:address2), with: values.address2
-      end
-
-      select values.country.name, from: Spree.t(:country) if values.country
-      fill_in Spree.t(:city), with: values.city
-      fill_in Spree.t(:zipcode), with: values.zipcode
-      select values.state.name, from: Spree.t(:state)
-      fill_in Spree.t(:phone), with: values.phone
-      fill_in Spree.t(:alternative_phone), with: values.alternative_phone if Spree::Config[:alternative_shipping_phone]
-    elsif values.is_a?(Hash)
-      values.each do |k, v|
-        if k == Spree.t(:country) || k == Spree.t(:state)
-          select v, from: k
-        else
-          fill_in k, with: v
-        end
-      end
-    else
-      raise "Invalid type #{values.class.name} for values"
-    end
-
-    expect(page).to have_css('#address_address_type') if type
-    if type == :user
-      select Spree.t(:user), from: Spree.t(:address_type)
-    elsif type == :bill
-      select Spree.t(:billing_address), from: Spree.t(:address_type)
-    elsif type == :ship
-      select Spree.t(:shipping_address), from: Spree.t(:address_type)
-    end
-  end
-
   # Expects the current path and query string to match the order customer
   # details path for an order, or the user account details path for a user.
   def expect_address_collection_path(order_or_user)
